backend/requirements.txt
fastapi
uvicorn
praw
anthropic
python-dotenv
pydantic
